requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-crispy-forms==2.0
django-phonenumber-field==7.0.2
django-widget-tweaks==1.4.12
idna==3.4
phonenumbers==8.13.7
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
python-decouple==3.8
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
twilio==7.17.0
tzdata==2022.7
urllib3==1.26.15
